Refactor and move CheckoutActivity to top package and rename the package CheckoutActivity to component
