+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/checkout/component/views/ActionCellContract.kt
@@ -1,4 +1,4 @@
-package com.karhoo.uisdk.screen.booking.checkout.checkoutActivity.views
+package com.karhoo.uisdk.screen.booking.checkout.component.views
 
 interface ActionCellContract {
     interface View {
