+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/checkout/component/views/ActionCellView.kt
@@ -1,4 +1,4 @@
-package com.karhoo.uisdk.screen.booking.checkout.checkoutActivity.views
+package com.karhoo.uisdk.screen.booking.checkout.component.views
 
 import android.content.Context
 import android.util.AttributeSet
