+++ uisdk/src/main/res/layout/uisdk_booking_checkout_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
 
-    <com.karhoo.uisdk.screen.booking.checkout.checkoutActivity.views.CheckoutView
+    <com.karhoo.uisdk.screen.booking.checkout.component.views.CheckoutView
         android:id="@+id/bookingCheckoutView"
         android:layout_width="match_parent"
         android:layout_height="0dp"
